requirements.txt
requests
pymongo
python-dotenv
flask
pandas 
matplotlib
